Add SOC die temperature and process utilization to report
Capture the SOC die temperature in the header of the report.
This uses the contents of /sys/class/thermal/thermal_zone0/temp
and calculates the temperature in degrees celsius.

Capture the process utilization and system load averages. This
requires the 'top' command to be installed in the system. We
can use this to gauge whether certain userspace or kernel
processes are contributing to excessive power consumption and
or system performance degradation.