Add support for gathering DPKG info
Add support for gathering dpkg info as that is the default package
manager used by TechNexion at this point.

Also, improve handling for cases where a package manager is present
but no packages are installed. For example, if the rpm package manager
is present, but no packages installed, the user should know that
no RPM packages were installed explicitly. Before there was just an
empty package list file, which could be confusing.